Iulian is right, if the MethodBinding is coming from a binary file, then we don't need to run it through the ScalaMethodVerifierProvider, because all the mixed-in methods have been added (therefore Java compilers sees the correct types).
